[refactor] #22 패키지 구조 만들면서 이름 변경된 엔티티 이름 패키지에 반영된 변경사항
